review & refactoring
